refactor(encodeActions): moved encodeActions to utils folder and replaced usage in useProposalCalls
