+++ apps/davi/src/stores/modules/1_5/fetchers/rpc/useProposalCalls/useProposalCalls.ts
@@ -5,22 +5,15 @@ import { useTheme } from 'styled-components';
 import { useTranslation } from 'react-i18next';
 
 import {
-  preventEmptyString,
-  ZERO_HASH,
   ERC20_APPROVE_SIGNATURE,
   getGuildOptionLabel,
+  encodeActions,
 } from 'utils';
 import { getBigNumberPercentage } from 'utils/bnPercentage';
 import { FetcherHooksInterface } from 'stores/types';
-import {
-  bulkDecodeCallsFromOptions,
-  decodeCall,
-} from 'hooks/Guilds/contracts/useDecodedCall';
+import { bulkDecodeCallsFromOptions } from 'hooks/Guilds/contracts/useDecodedCall';
 import useProposalMetadata from 'hooks/Guilds/useProposalMetadata';
-import {
-  RichContractData,
-  useRichContractRegistry,
-} from 'hooks/Guilds/contracts/useRichContractRegistry';
+import { useRichContractRegistry } from 'hooks/Guilds/contracts/useRichContractRegistry';
 import { Call, Option } from 'components/ActionsBuilder/types';
 import { EMPTY_CALL } from 'Modules/Guilds/pages/CreateProposal';
 import { useTotalLocked } from '../../subgraph';
@@ -28,44 +21,6 @@ import { useTotalLocked } from '../../subgraph';
 const isApprovalData = (data: string) =>
   data && data?.substring(0, 10) === ERC20_APPROVE_SIGNATURE;
 const isApprovalCall = (call: Call) => isApprovalData(call?.data);
-const isZeroHash = (data: string) => data === ZERO_HASH;
-
-const encodeActions = async (
-  calls: Call[],
-  contracts: RichContractData[],
-  chainId: number
-) => {
-  const filteredCalls = calls.filter(
-    call => !isZeroHash(call?.data) || !preventEmptyString(call?.value).isZero()
-  );
-
-  const encodedActions = await Promise.all(
-    filteredCalls.map(async (call: Call) => {
-      if (!!call?.approvalCall) {
-        // If current call is an "spending" call will have a inner approvalCall
-        const { decodedCall: decodedApprovalCall } = await decodeCall(
-          call?.approvalCall,
-          contracts,
-          chainId
-        );
-        // Avoid spreading unnecesary approvalCall;
-        const { approvalCall, ...newCall } = call;
-
-        return {
-          ...newCall,
-          approval: {
-            ...decodedApprovalCall,
-            amount: decodedApprovalCall?.args?._value,
-            token: decodedApprovalCall?.to,
-          },
-        };
-      }
-      return call;
-    })
-  );
-
-  return encodedActions;
-};
 
 type IUseProposalCalls = FetcherHooksInterface['useProposalCalls'];
